requirements.txt
# Football Analysis - Requirements
# Installation: pip install -r requirements.txt
# NOTE: NumPy 1.26.4 must be used due to compatibility issues with NumPy 2.0

# ============================================
# CRITICAL LIBRARIES
# ============================================

# NumPy - Use 1.26.4 for PyTorch and OpenCV compatibility (Do not use 2.0!)
numpy==1.26.4

# Computer Vision & Image Processing
opencv-python-headless==4.11.0.86
Pillow>=10.0.0

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Object Detection & Tracking
ultralytics>=8.0.0
supervision==0.27.0

# Sports Analysis
pytorchvideo>=0.1.5

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0
umap-learn>=0.5.0
transformers>=4.30.0

# Data Processing
pandas>=2.0.0
h5py==3.14.0

# Video Processing
decord==0.6.0

# Data Augmentation
albumentations==2.0.8

# Utilities
tqdm>=4.66.0
matplotlib>=3.7.0
requests>=2.31.0
gdown>=4.7.0
psutil>=5.9.0
tensorboard>=2.14.0

# ============================================
# GUI & DISPLAY
# ============================================
PyQt5>=5.15.0

# ============================================
# EXTRA DEPENDENCIES (Optional)
# ============================================
lmdb>=1.4.0
nltk>=3.8.0
editdistance>=0.6.0
timm>=0.9.0
pytorch_lightning>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pyclipper>=1.3.0
shapely>=2.0.0
torch_geometric>=2.3.0
fvcore>=0.1.5
iopath>=0.1.9
termcolor>=2.3.0
yacs>=0.1.8

# ============================================
# GIT PACKAGES (Manual installation required)
# ============================================
# git+https://github.com/roboflow/sports.git
